requirements.txt
kaggle==1.5.6
matplotlib==3.1.3
newspaper3k==0.2.8
numpy==1.18.1
scikit-learn==0.22.1
streamlit==0.63.1